wip: created perun class, created singleton decorator, removed backend decorator
