+++ perun/__main__.py
@@ -1,5 +1,5 @@
 """perun module."""
 if __name__ == "__main__":
-    from perun.cli import cli
+    from perun.api.cli import cli
 
     cli()
