+++ perun/backend/__init__.py
@@ -1,6 +1,6 @@
 """Backend module."""
 # flake8: noqa
-from .backend import Backend, backend, backends
+from .backend import Backend
 from .intel_rapl import IntelRAPLBackend
 from .nvml import NVMLBackend
 from .psutil import PSUTILBackend
