+++ perun/util.py
@@ -1,14 +1,21 @@
 """Util module."""
 import os
-import platform
-from datetime import datetime, timedelta
+from datetime import datetime
 from pathlib import Path
-from typing import Any, Callable, Dict, Set, Tuple, Union
+from typing import Callable, Union
 
-import numpy as np
+from perun import config
 
-from perun import config, log
-from perun.data_model.measurement_type import Magnitude, MetricMetaData, Unit
+
+def singleton(class_):
+    instances = {}
+
+    def getinstance(*args, **kwargs):
+        if class_ not in instances:
+            instances[class_] = class_(*args, **kwargs)
+        return instances[class_]
+
+    return getinstance
 
 
 def getRunName(app: Union[Path, Callable]) -> str:
@@ -52,78 +59,3 @@ def getRunId(startime: datetime) -> str:
         return os.environ["SLURM_JOB_ID"]
     else:
         return startime.isoformat()
-
-
-def value2str(
-    value: np.number, metric_md: MetricMetaData
-) -> Tuple[str, float, Magnitude]:
-    """Return a printable representation of the value based on its metric metadata. A printable value should not have more than 3 digits after before the decimal comma/dot.
-
-    Args:
-        value (np.number): Value to format.
-        metric_md (MetricMetaData): Value metadata.
-
-    Returns:
-        Tuple[str, float, Magnitude]: Formated value, transformation factor used, and the new magnitude prefix.
-    """
-    if metric_md.unit == Unit.WATT or metric_md.unit == Unit.JOULE:
-        transformFactor = 1
-        for mag in reversed(Magnitude):
-            if value > mag.value:
-                transformFactor = mag.value
-                break
-
-        newValue = value / transformFactor
-        newMag = Magnitude(metric_md.mag.value * transformFactor)
-        return f"{newValue:.3f}", transformFactor, newMag
-
-    elif metric_md.unit == Unit.PERCENT:
-        return f"{value:.3f}", 1.0, metric_md.mag
-    elif metric_md.unit == Unit.SECOND:
-        return str(timedelta(seconds=float(value))), 1.0, Magnitude.ONE
-    elif metric_md.unit == Unit.BYTE:
-        transformFactor = 1
-        newMag = Magnitude.ONE
-        for magFactor, m in zip(
-            [1024**3, 1024**2, 1024**1],
-            [Magnitude.GIGA, Magnitude.MEGA, Magnitude.KILO],
-        ):
-            if value > magFactor:
-                transformFactor = magFactor
-                newMag = m
-                break
-
-        newValue = value / transformFactor
-        return f"{newValue:.3f}", transformFactor, newMag
-    else:
-        return f"{value:.3f}", 1.0, metric_md.mag
-
-
-def getHostMetadata(backendConfig: Dict[str, Set[str]]) -> Dict[str, Any]:
-    """Return dictionary with the full system metadata based on the provided backend configuration.
-
-    :param backendConfig: Sensor backend configuration to include in the metadata object.
-    :type backendConfig: Dict[str, Set[str]]
-    :return: Dictionary with system metadata
-    :rtype: Dict[str, Any]
-    """
-    from perun.backend import backends
-
-    metadata = {}
-    for name, method in platform.__dict__.items():
-        if callable(method):
-            try:
-                metadata[name] = method()
-            except Exception as e:
-                log.warn(f"platform method {name} did not work")
-                log.warn(e)
-
-    metadata["backends"] = {}
-    for backend in backends:
-        if backend.name in backendConfig:
-            metadata["backends"][backend.name] = {}
-            sensors = backend.getSensors(backendConfig[backend.name])
-            for sensor in sensors:
-                metadata["backends"][backend.name][sensor.id] = sensor.metadata
-
-    return metadata
