+++ tests/perun/test_cli.py
@@ -3,7 +3,7 @@ from configparser import ConfigParser
 from click.testing import CliRunner, Result
 
 import perun
-from perun.cli import cli
+from perun.api.cli import cli
 from perun.configuration import _default_config
